@@ -49,36 +51,53 @@
   },
   "homepage": "https://github.com/apollostack/graphql-tools#readme",
   "dependencies": {
-    "apollo-link": "^1.2.3",
-    "apollo-utilities": "^1.0.1",
+    "apollo-link": "^1.2.13",
+    "apollo-link-http-common": "^0.2.15",
     "deprecated-decorator": "^0.1.6",
-    "iterall": "^1.1.3",
-    "uuid": "^3.1.0"
+    "extract-files": "^7.0.0",
+    "form-data": "^3.0.0",
+    "iterall": "^1.3.0",
+    "node-fetch": "^2.6.0",
+    "uuid": "^7.0.2"
   },
   "peerDependencies": {
-    "graphql": "^0.13.0 || ^14.0.0"
+    "graphql": "^0.12.0 || ^0.13.0 || ^14.0.0 || ^15.0.0-rc"
   },
   "devDependencies": {
-    "@types/chai": "4.0.10",
-    "@types/dateformat": "^1.0.1",
-    "@types/mocha": "^2.2.44",
-    "@types/node": "^8.0.47",
-    "@types/uuid": "^3.4.3",
-    "@types/zen-observable": "^0.5.3",
-    "body-parser": "^1.18.2",
-    "chai": "^4.1.2",
+    "@types/chai": "^4.2.11",
+    "@types/dateformat": "^3.0.1",
+    "@types/express": "^4.17.3",
+    "@types/extract-files": "^3.1.0",
+    "@types/graphql-type-json": "^0.3.2",
+    "@types/graphql-upload": "^8.0.3",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^13.9.3",
+    "@types/node-fetch": "^2.5.5",
+    "@types/uuid": "^7.0.2",
+    "@typescript-eslint/eslint-plugin": "^2.25.0",
+    "@typescript-eslint/parser": "^2.25.0",
+    "babel-eslint": "^10.1.0",
+    "body-parser": "^1.19.0",
+    "chai": "^4.2.0",
+    "coveralls": "^3.0.11",
+    "dataloader": "^2.0.0",
     "dateformat": "^3.0.3",
-    "express": "^4.16.2",
-    "graphql": "^14.5.8",
-    "graphql-subscriptions": "^1.0.0",
-    "graphql-type-json": "^0.1.4",
-    "istanbul": "^0.4.5",
-    "mocha": "^4.0.1",
-    "prettier": "^1.7.4",
-    "remap-istanbul": "0.9.6",
-    "rimraf": "^2.6.2",
-    "source-map-support": "^0.5.0",
-    "tslint": "^5.8.0",
-    "typescript": "^3.6.4"
+    "eslint": "^6.8.0",
+    "eslint-plugin-import": "2.20.1",
+    "eslint-watch": "^6.0.1",
+    "express": "^4.17.1",
+    "express-graphql": "^0.9.0",
+    "graphql": "^14.6.0",
+    "graphql-subscriptions": "^1.1.0",
+    "graphql-type-json": "^0.3.1",
+    "graphql-upload": "^10.0.0",
+    "mocha": "^7.1.1",
+    "nyc": "^15.0.0",
+    "prettier": "^2.0.2",
+    "rimraf": "^3.0.2",
+    "source-map-support": "^0.5.16",
+    "standard-version": "^7.1.0",
+    "typescript": "^3.8.3",
+    "zen-observable-ts": "^0.8.20"
   }
 }
