@@ -1,6 +1,10 @@
 export * from './Interfaces';
-export * from './makeExecutableSchema';
+export * from './delegate';
+export * from './generate';
+export * from './links';
 export * from './mock';
-export * from './stitching';
-export * from './transforms';
-export { SchemaDirectiveVisitor } from './schemaVisitor';
+export * from './polyfills';
+export * from './scalars';
+export * from './stitch';
+export * from './wrap';
+export * from './utils';
