@@ -1,19 +1,23 @@
-/* tslint:disable:no-unused-expression */
-
+import { GraphQLSchema, graphql } from 'graphql';
 import { expect } from 'chai';
-import {
-  GraphQLSchema,
-  graphql
-} from 'graphql';
-import { propertySchema, bookingSchema, sampleData, Property } from './testingSchemas';
-import delegateToSchema from '../stitching/delegateToSchema';
-import mergeSchemas from '../stitching/mergeSchemas';
-import { IResolvers } from '../Interfaces';
 
-function findPropertyByLocationName (
+import delegateToSchema from '../delegate/delegateToSchema';
+import mergeSchemas from '../stitch/mergeSchemas';
+import { IResolvers } from '../Interfaces';
+import { makeExecutableSchema } from '../generate';
+import { wrapSchema } from '../wrap';
+
+import {
+  propertySchema,
+  bookingSchema,
+  sampleData,
+  Property,
+} from './testingSchemas';
+
+function findPropertyByLocationName(
   properties: { [key: string]: Property },
-  name: string
-): Property {
+  name: string,
+): Property | undefined {
   for (const key of Object.keys(properties)) {
     const property = properties[key];
     if (property.location.name === name) {
@@ -34,35 +38,37 @@ const COORDINATES_QUERY = `
   }
 `;
 
-function proxyResolvers (spec: string): IResolvers {
+function proxyResolvers(spec: string): IResolvers {
   return {
     Booking: {
       property: {
         fragment: '... on Booking { propertyId }',
-        resolve (booking, args, context, info) {
-          const delegateFn = spec === 'standalone' ? delegateToSchema :
-            info.mergeInfo.delegateToSchema;
-          return delegateFn({
+        resolve(booking, _args, context, info) {
+          const delegateFn =
+            spec === 'standalone'
+              ? delegateToSchema
+              : info.mergeInfo.delegateToSchema;
+          return delegateFn?.({
             schema: propertySchema,
             operation: 'query',
             fieldName: 'propertyById',
             args: { id: booking.propertyId },
             context,
-            info
+            info,
           });
-        }
-      }
+        },
+      },
     },
     Location: {
       coordinates: {
         fragment: '... on Location { name }',
-        resolve (location, args, context, info) {
+        resolve: (location) => {
           const name = location.name;
-          return findPropertyByLocationName(sampleData.Property, name)
-            .location.coordinates;
-        }
-      }
-    }
+          return findPropertyByLocationName(sampleData.Property, name).location
+            .coordinates;
+        },
+      },
+    },
   };
 }
 
@@ -77,32 +83,131 @@ const proxyTypeDefs = `
 
 describe('stitching', () => {
   describe('delegateToSchema', () => {
-    ['standalone', 'info.mergeInfo'].forEach(spec => {
-      context(spec, () => {
+    ['standalone', 'info.mergeInfo'].forEach((spec) => {
+      describe(spec, () => {
         let schema: GraphQLSchema;
         before(() => {
           schema = mergeSchemas({
             schemas: [bookingSchema, propertySchema, proxyTypeDefs],
-            resolvers: proxyResolvers(spec)
+            resolvers: proxyResolvers(spec),
           });
         });
         it('should add fragments for deep types', async () => {
-          const result = await graphql(schema, COORDINATES_QUERY,
-            {}, {}, { bookingId: 'b1' });
+          const result = await graphql(
+            schema,
+            COORDINATES_QUERY,
+            {},
+            {},
+            { bookingId: 'b1' },
+          );
 
           expect(result).to.deep.equal({
             data: {
               bookingById: {
                 property: {
                   location: {
-                    coordinates: sampleData.Property.p1.location.coordinates
-                  }
-                }
-              }
-            }
+                    coordinates: sampleData.Property.p1.location.coordinates,
+                  },
+                },
+              },
+            },
           });
         });
       });
+    });
+  });
+});
+
+describe('schema delegation', () => {
+  it('should work even when there are default fields', async () => {
+    const schema = makeExecutableSchema({
+      typeDefs: `
+        scalar JSON
+        type Data {
+          json(input: JSON = "test"): JSON
+        }
+        type Query {
+          data: Data
+        }
+      `,
+      resolvers: {
+        Query: {
+          data: () => ({}),
+        },
+        Data: {
+          json: (_root, args, context, info) =>
+            delegateToSchema({
+              schema: propertySchema,
+              fieldName: 'jsonTest',
+              args,
+              context,
+              info,
+            }),
+        },
+      },
+    });
+
+    const result = await graphql(
+      schema,
+      `
+        query {
+          data {
+            json
+          }
+        }
+      `,
+    );
+
+    expect(result).to.deep.equal({
+      data: {
+        data: {
+          json: 'test',
+        },
+      },
+    });
+  });
+
+  it('should work even with variables', async () => {
+    const innerSchema = makeExecutableSchema({
+      typeDefs: `
+        type User {
+          id(show: Boolean): ID
+        }
+        type Query {
+          user: User
+        }
+      `,
+      resolvers: {
+        Query: {
+          user: () => ({}),
+        },
+        User: {
+          id: () => '123',
+        },
+      },
+    });
+    const schema = wrapSchema(innerSchema);
+
+    const result = await graphql(
+      schema,
+      `
+        query($show: Boolean) {
+          user {
+            id(show: $show)
+          }
+        }
+      `,
+      null,
+      null,
+      { show: true },
+    );
+
+    expect(result).to.deep.equal({
+      data: {
+        user: {
+          id: '123',
+        },
+      },
     });
   });
 });
