@@ -1,12 +1,18 @@
 import { expect } from 'chai';
-import { graphql, GraphQLResolveInfo } from 'graphql';
-import { addMockFunctionsToSchema, MockList, mockServer } from '../mock';
+import {
+  graphql,
+  GraphQLResolveInfo,
+  GraphQLSchema,
+  GraphQLFieldResolver,
+} from 'graphql';
+
+import { addMocksToSchema, MockList, mockServer } from '../mock/index';
 import {
   buildSchemaFromTypeDefinitions,
-  addResolveFunctionsToSchema,
+  addResolversToSchema,
   makeExecutableSchema,
-} from '../makeExecutableSchema';
-import 'mocha';
+} from '../generate/index';
+import { IMocks } from '../Interfaces';
 
 describe('Mock', () => {
   const shorthand = `
@@ -72,33 +78,34 @@ describe('Mock', () => {
 
   const resolveFunctions = {
     BirdsAndBees: {
-      __resolveType(data: any, context: any, info: GraphQLResolveInfo) {
+      __resolveType(data: any, _context: any, info: GraphQLResolveInfo) {
         return info.schema.getType(data.__typename);
       },
     },
     Flying: {
-      __resolveType(data: any, context: any, info: GraphQLResolveInfo) {
+      __resolveType(data: any, _context: any, info: GraphQLResolveInfo) {
         return info.schema.getType(data.__typename);
       },
     },
   };
 
   it('throws an error if you forget to pass schema', () => {
-    expect(() => (<any>addMockFunctionsToSchema)({})).to.throw(
-      'Must provide schema to mock',
-    );
+    expect(() => addMocksToSchema({})).to.throw('Must provide schema to mock');
   });
 
   it('throws an error if the property "schema" on the first argument is not of type GraphQLSchema', () => {
-    expect(() => (<any>addMockFunctionsToSchema)({ schema: {} })).to.throw(
-      'Value at "schema" must be of type GraphQLSchema',
-    );
+    expect(() =>
+      addMocksToSchema({ schema: ({} as unknown) as GraphQLSchema }),
+    ).to.throw('Value at "schema" must be of type GraphQLSchema');
   });
 
   it('throws an error if second argument is not a Map', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     expect(() =>
-      (<any>addMockFunctionsToSchema)({ schema: jsSchema, mocks: ['a'] }),
+      addMocksToSchema({
+        schema: jsSchema,
+        mocks: (['a'] as unknown) as IMocks,
+      }),
     ).to.throw('mocks must be of type Object');
   });
 
@@ -106,14 +113,17 @@ describe('Mock', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { Int: 55 };
     expect(() =>
-      (<any>addMockFunctionsToSchema)({ schema: jsSchema, mocks: mockMap }),
+      addMocksToSchema({
+        schema: jsSchema,
+        mocks: (mockMap as unknown) as IMocks,
+      }),
     ).to.throw('mockFunctionMap[Int] must be a function');
   });
 
   it('mocks the default types for you', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = {};
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnInt
       returnFloat
@@ -121,7 +131,7 @@ describe('Mock', () => {
       returnString
       returnID
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnInt']).to.be.within(-1000, 1000);
       expect(res.data['returnFloat']).to.be.within(-1000, 1000);
       expect(res.data['returnBoolean']).to.be.a('boolean');
@@ -157,9 +167,7 @@ describe('Mock', () => {
       .query(testQuery)
       .then((res: any) => {
         expect(res.data.returnInt).to.equal(12345);
-        expect(res.data.returnFloat)
-          .to.be.a('number')
-          .within(-1000, 1000);
+        expect(res.data.returnFloat).to.be.a('number').within(-1000, 1000);
         expect(res.data.returnBoolean).to.be.a('boolean');
         expect(res.data.returnString).to.be.a('string');
         expect(res.data.returnID).to.be.a('string');
@@ -177,7 +185,7 @@ describe('Mock', () => {
         returnString: () => 'someString',
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
+    addResolversToSchema(jsSchema, resolvers);
     const testQuery = `{
       returnInt
       returnString
@@ -235,9 +243,7 @@ describe('Mock', () => {
       .query(testQuery)
       .then((res: any) => {
         expect(res.data.returnInt).to.equal(12345);
-        expect(res.data.returnFloat)
-          .to.be.a('number')
-          .within(-1000, 1000);
+        expect(res.data.returnFloat).to.be.a('number').within(-1000, 1000);
         expect(res.data.returnBoolean).to.be.a('boolean');
         expect(res.data.returnString).to.be.a('string');
         expect(res.data.returnID).to.be.a('string');
@@ -253,14 +259,14 @@ describe('Mock', () => {
     let spy = 0;
     const resolvers = {
       BirdsAndBees: {
-        __resolveType(data: any, context: any, info: GraphQLResolveInfo) {
+        __resolveType(data: any, _context: any, info: GraphQLResolveInfo) {
           ++spy;
           return info.schema.getType(data.__typename);
         },
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
-    addMockFunctionsToSchema({
+    addResolversToSchema(jsSchema, resolvers);
+    addMocksToSchema({
       schema: jsSchema,
       mocks: {},
       preserveResolvers: true,
@@ -277,7 +283,7 @@ describe('Mock', () => {
         }
       }
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((_res) => {
       // the resolveType has been called twice
       expect(spy).to.equal(2);
     });
@@ -287,18 +293,18 @@ describe('Mock', () => {
   it('can mock Enum', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = {};
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnEnum
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnEnum']).to.be.oneOf(['A', 'B', 'C']);
     });
   });
 
   it('can mock Unions', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
-    addResolveFunctionsToSchema(jsSchema, resolveFunctions);
+    addResolversToSchema(jsSchema, resolveFunctions);
     const mockMap = {
       Int: () => 10,
       String: () => 'aha',
@@ -307,7 +313,7 @@ describe('Mock', () => {
         returnBirdsAndBees: () => new MockList(40),
       }),
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnBirdsAndBees {
         ... on Bird {
@@ -320,7 +326,7 @@ describe('Mock', () => {
         }
       }
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       // XXX this test is expected to fail once every 2^40 times ;-)
       expect(res.data['returnBirdsAndBees']).to.deep.include({
         returnInt: 10,
@@ -335,7 +341,7 @@ describe('Mock', () => {
 
   it('can mock Interfaces by default', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
-    addResolveFunctionsToSchema(jsSchema, resolveFunctions);
+    addResolversToSchema(jsSchema, resolveFunctions);
     const mockMap = {
       Int: () => 10,
       String: () => 'aha',
@@ -344,7 +350,7 @@ describe('Mock', () => {
         returnFlying: () => new MockList(40),
       }),
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -361,7 +367,7 @@ describe('Mock', () => {
         }
       }
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnFlying']).to.deep.include({
         returnInt: 10,
         returnString: 'aha',
@@ -375,27 +381,28 @@ describe('Mock', () => {
 
   it('can support explicit Interface mock', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
-    addResolveFunctionsToSchema(jsSchema, resolveFunctions);
+    addResolversToSchema(jsSchema, resolveFunctions);
     let spy = 0;
     const mockMap = {
-      Bird: (root: any, args: any) => ({
+      Bird: (_root: any, args: any) => ({
         id: args.id,
         returnInt: 100,
       }),
-      Bee: (root: any, args: any) => ({
+      Bee: (_root: any, args: any) => ({
         id: args.id,
         returnInt: 200,
       }),
-      Flying: (root: any, args: any) => {
+      Flying: (_root: any, args: any) => {
         spy++;
         const { id } = args;
         const type = id.split(':')[0];
-        // tslint:disable-next-line
-        const __typename = ['Bird', 'Bee'].find(r => r.toLowerCase() === type);
+        const __typename = ['Bird', 'Bee'].find(
+          (r) => r.toLowerCase() === type,
+        );
         return { __typename };
       },
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -407,7 +414,7 @@ describe('Mock', () => {
       }
     }`;
 
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(spy).to.equal(1); // to make sure that Flying possible types are not randomly selected
       expect(res.data['node']).to.include({
         id: 'bee:123456',
@@ -418,27 +425,27 @@ describe('Mock', () => {
 
   it('can support explicit UnionType mock', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
-    addResolveFunctionsToSchema(jsSchema, resolveFunctions);
+    addResolversToSchema(jsSchema, resolveFunctions);
     let spy = 0;
     const mockMap = {
-      Bird: (root: any, args: any) => ({
+      Bird: (_root: any, args: any) => ({
         id: args.id,
         returnInt: 100,
       }),
-      Bee: (root: any, args: any) => ({
+      Bee: (_root: any, args: any) => ({
         id: args.id,
         returnEnum: 'A',
       }),
-      BirdsAndBees: (root: any, args: any) => {
+      BirdsAndBees: (_root: any, args: any) => {
         spy++;
         const { id } = args;
         const type = id.split(':')[0];
         return {
-          __typename: ['Bird', 'Bee'].find(r => r.toLowerCase() === type),
+          __typename: ['Bird', 'Bee'].find((r) => r.toLowerCase() === type),
         };
       },
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -452,7 +459,7 @@ describe('Mock', () => {
         }
       }`;
 
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(spy).to.equal(1);
       expect(res.data['node2']).to.include({
         id: 'bee:123456',
@@ -463,21 +470,19 @@ describe('Mock', () => {
 
   it('throws an error when __typename is not returned within an explicit interface mock', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
-    addResolveFunctionsToSchema(jsSchema, resolveFunctions);
+    addResolversToSchema(jsSchema, resolveFunctions);
     const mockMap = {
-      Bird: (root: any, args: any) => ({
+      Bird: (_root: any, args: any) => ({
         id: args.id,
         returnInt: 100,
       }),
-      Bee: (root: any, args: any) => ({
+      Bee: (_root: any, args: any) => ({
         id: args.id,
         returnInt: 100,
       }),
-      Flying: (root: any, args: any): void => {
-        return;
-      },
+      Flying: (_root: any, _args: any) => ({}),
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
         node(id:"bee:123456"){
           id,
@@ -485,21 +490,21 @@ describe('Mock', () => {
         }
       }`;
     const expected = 'Please return a __typename in "Flying"';
-    return graphql(jsSchema, testQuery).then(res => {
-      expect((<any>res.errors[0]).originalError.message).to.equal(expected);
+    return graphql(jsSchema, testQuery).then((res) => {
+      expect(res.errors[0].originalError.message).to.equal(expected);
     });
   });
 
   it('throws an error in resolve if mock type is not defined', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = {};
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnMockError
     }`;
     const expected = 'No mock defined for type "MissingMockType"';
-    return graphql(jsSchema, testQuery).then(res => {
-      expect((<any>res.errors[0]).originalError.message).to.equal(expected);
+    return graphql(jsSchema, testQuery).then((res) => {
+      expect(res.errors[0].originalError.message).to.equal(expected);
     });
   });
 
@@ -512,13 +517,13 @@ describe('Mock', () => {
         __parseLiteral: (val: string) => val,
       },
       RootQuery: {
-        returnMockError: () => <string>undefined,
+        returnMockError: (): string => undefined,
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
+    addResolversToSchema(jsSchema, resolvers);
 
     const mockMap = {};
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -527,8 +532,8 @@ describe('Mock', () => {
       returnMockError
     }`;
     const expected = 'No mock defined for type "MissingMockType"';
-    return graphql(jsSchema, testQuery).then(res => {
-      expect((<any>res.errors[0]).originalError.message).to.equal(expected);
+    return graphql(jsSchema, testQuery).then((res) => {
+      expect(res.errors[0].originalError.message).to.equal(expected);
     });
   });
 
@@ -544,10 +549,10 @@ describe('Mock', () => {
         returnMockError: () => '10-11-2012',
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
+    addResolversToSchema(jsSchema, resolvers);
 
     const mockMap = {};
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -558,7 +563,7 @@ describe('Mock', () => {
     const expected = {
       returnMockError: '10-11-2012',
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
       expect(res.errors).to.equal(undefined);
     });
@@ -567,11 +572,11 @@ describe('Mock', () => {
   it('can mock an Int', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { Int: () => 55 };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnInt
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnInt']).to.equal(55);
     });
   });
@@ -579,77 +584,77 @@ describe('Mock', () => {
   it('can mock a Float', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { Float: () => 55.5 };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnFloat
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnFloat']).to.equal(55.5);
     });
   });
   it('can mock a String', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { String: () => 'a string' };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnString
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnString']).to.equal('a string');
     });
   });
   it('can mock a Boolean', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { Boolean: () => true };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnBoolean
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnBoolean']).to.equal(true);
     });
   });
   it('can mock an ID', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { ID: () => 'ea5bdc19' };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnID
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnID']).to.equal('ea5bdc19');
     });
   });
   it('nullable type is nullable', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { String: (): null => null };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnNullableString
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnNullableString']).to.equal(null);
     });
   });
   it('can mock a nonNull type', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { String: () => 'nonnull' };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnNonNullString
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnNonNullString']).to.equal('nonnull');
     });
   });
   it('nonNull type is not nullable', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { String: (): null => null };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnNonNullString
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.equal(null);
       expect(res.errors.length).to.equal(1);
     });
@@ -660,14 +665,14 @@ describe('Mock', () => {
       String: () => 'abc',
       Int: () => 123,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnObject { returnInt, returnString }
     }`;
     const expected = {
       returnObject: { returnInt: 123, returnString: 'abc' },
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -675,14 +680,14 @@ describe('Mock', () => {
   it('can mock a list of ints', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = { Int: () => 123 };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnListOfInt
     }`;
     const expected = {
       returnListOfInt: [123, 123],
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -693,7 +698,7 @@ describe('Mock', () => {
       String: () => 'a',
       Int: () => 1,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnListOfListOfObject { returnInt, returnString }
     }`;
@@ -709,18 +714,18 @@ describe('Mock', () => {
         ],
       ],
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
 
-  it('does not mask resolve functions if you tell it not to', () => {
+  it('does not mask resolvers if you tell it not to', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = {
       RootQuery: () => ({
-        returnInt: (root: any, args: { [key: string]: any }) => 42, // a) in resolvers, will not be used
-        returnFloat: (root: any, args: { [key: string]: any }) => 1.3, // b) not in resolvers, will be used
-        returnString: (root: any, args: { [key: string]: any }) =>
+        returnInt: (_root: any, _args: { [key: string]: any }) => 42, // a) in resolvers, will not be used
+        returnFloat: (_root: any, _args: { [key: string]: any }) => 1.3, // b) not in resolvers, will be used
+        returnString: (_root: any, _args: { [key: string]: any }) =>
           Promise.resolve('foo'), // c) in resolvers, will not be used
       }),
     };
@@ -730,8 +735,8 @@ describe('Mock', () => {
         returnString: () => Promise.resolve('bar'), // see c)
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
-    addMockFunctionsToSchema({
+    addResolversToSchema(jsSchema, resolvers);
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -746,7 +751,7 @@ describe('Mock', () => {
       returnFloat: 1.3, // b) from mock
       returnString: 'bar', // c) from resolvers, not masked by mock (and promise)
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -760,7 +765,7 @@ describe('Mock', () => {
       }),
       Int: () => 15,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnObject{
         returnInt
@@ -775,7 +780,7 @@ describe('Mock', () => {
       },
       returnInt: 15,
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -791,7 +796,7 @@ describe('Mock', () => {
         }),
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
+    addResolversToSchema(jsSchema, resolvers);
     const mockMap = {
       returnListOfInt: () => [5, 6, 7],
       Bird: () => ({
@@ -799,7 +804,7 @@ describe('Mock', () => {
         returnString: 'woot!?', // b) another part of a Bird
       }),
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -818,7 +823,7 @@ describe('Mock', () => {
         returnString: 'woot!?', // from the mock, see b)
       },
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -834,14 +839,14 @@ describe('Mock', () => {
           }),
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
+    addResolversToSchema(jsSchema, resolvers);
     const mockMap = {
       Bird: () => ({
         returnInt: 3, // see a)
         returnString: 'woot!?', // b) another part of a Bird
       }),
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -858,7 +863,7 @@ describe('Mock', () => {
         returnString: 'woot!?', // from the mock, see b)
       },
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -876,14 +881,14 @@ describe('Mock', () => {
         returnObject: () => objProxy,
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
+    addResolversToSchema(jsSchema, resolvers);
     const mockMap = {
       Bird: () => ({
         returnInt: 3, // see a)
         returnString: 'woot!?', // b) another part of a Bird
       }),
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -900,7 +905,7 @@ describe('Mock', () => {
         returnString: 'woot!?', // from the mock, see b)
       },
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -924,7 +929,7 @@ describe('Mock', () => {
     const mockMap = {
       Int: () => 123, // b) mock of Int.
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -943,7 +948,7 @@ describe('Mock', () => {
       },
       returnString: 'woot!?', // from the mock, see a)
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -969,7 +974,7 @@ describe('Mock', () => {
     const mockMap = {
       Int: () => 123, // b) mock of Int.
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -988,7 +993,7 @@ describe('Mock', () => {
       },
       returnString: 'woot!?', // from the mock, see a)
     };
-    return graphql(jsSchema, testQuery, undefined, {}).then(res => {
+    return graphql(jsSchema, testQuery, undefined, {}).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1015,7 +1020,7 @@ describe('Mock', () => {
     const mockMap = {
       Int: () => 123, // b) mock of Int.
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -1034,7 +1039,7 @@ describe('Mock', () => {
       },
       returnString: 'woot!?', // from the mock, see a)
     };
-    return graphql(jsSchema, testQuery, undefined, {}).then(res => {
+    return graphql(jsSchema, testQuery, undefined, {}).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1046,11 +1051,11 @@ describe('Mock', () => {
         returnString: (): string => null, // a) resolve of a string
       },
     };
-    addResolveFunctionsToSchema(jsSchema, resolvers);
+    addResolversToSchema(jsSchema, resolvers);
     const mockMap = {
       Int: () => 666, // b) mock of Int.
     };
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema: jsSchema,
       mocks: mockMap,
       preserveResolvers: true,
@@ -1067,9 +1072,9 @@ describe('Mock', () => {
         returnInt: 666, // from the mock, see b)
         returnString: 'Hello World', // from mock default values.
       },
-      returnString: null as string, /// from the mock, see a)
+      returnString: null as string, // from the mock, see a)
     };
-    return graphql(jsSchema, testQuery, undefined, {}).then(res => {
+    return graphql(jsSchema, testQuery, undefined, {}).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1078,17 +1083,17 @@ describe('Mock', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = {
       RootQuery: () => ({
-        returnStringArgument: (o: any, a: { [key: string]: any }) => a['s'],
+        returnStringArgument: (_o: any, a: { [key: string]: any }) => a['s'],
       }),
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnStringArgument(s: "adieu")
     }`;
     const expected = {
       returnStringArgument: 'adieu',
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1097,17 +1102,17 @@ describe('Mock', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = {
       RootMutation: () => ({
-        returnStringArgument: (o: any, a: { [key: string]: any }) => a['s'],
+        returnStringArgument: (_o: any, a: { [key: string]: any }) => a['s'],
       }),
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `mutation {
       returnStringArgument(s: "adieu")
     }`;
     const expected = {
       returnStringArgument: 'adieu',
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1118,14 +1123,14 @@ describe('Mock', () => {
       RootQuery: () => ({ returnListOfInt: () => new MockList(3) }),
       Int: () => 12,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnListOfInt
     }`;
     const expected = {
       returnListOfInt: [12, 12, 12],
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1136,11 +1141,11 @@ describe('Mock', () => {
       RootQuery: () => ({ returnListOfInt: () => new MockList([10, 20]) }),
       Int: () => 12,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnListOfInt
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['returnListOfInt']).to.have.length.within(10, 20);
       expect(res.data['returnListOfInt'][0]).to.equal(12);
     });
@@ -1150,17 +1155,17 @@ describe('Mock', () => {
     const jsSchema = buildSchemaFromTypeDefinitions(shorthand);
     const mockMap = {
       RootQuery: () => ({
-        returnListOfIntArg: (o: any, a: { [key: string]: any }) =>
+        returnListOfIntArg: (_o: any, a: { [key: string]: any }) =>
           new MockList(a['l']),
       }),
       Int: () => 12,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       l3: returnListOfIntArg(l: 3)
       l5: returnListOfIntArg(l: 5)
     }`;
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data['l3'].length).to.equal(3);
       expect(res.data['l5'].length).to.equal(5);
     });
@@ -1174,22 +1179,23 @@ describe('Mock', () => {
       }),
       Int: () => 12,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnListOfInt
     }`;
     const expected = {
       returnListOfInt: [33, 33],
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
 
   it('throws an error if the second argument to MockList is not a function', () => {
-    expect(() => new (<any>MockList)(5, 'abc')).to.throw(
-      'Second argument to MockList must be a function or undefined',
-    );
+    expect(
+      () =>
+        new MockList(5, ('abc' as unknown) as GraphQLFieldResolver<any, any>),
+    ).to.throw('Second argument to MockList must be a function or undefined');
   });
 
   it('lets you nest MockList in MockList', () => {
@@ -1200,14 +1206,17 @@ describe('Mock', () => {
       }),
       Int: () => 12,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnListOfListOfInt
     }`;
     const expected = {
-      returnListOfListOfInt: [[12, 12, 12], [12, 12, 12]],
+      returnListOfListOfInt: [
+        [12, 12, 12],
+        [12, 12, 12],
+      ],
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1219,19 +1228,19 @@ describe('Mock', () => {
         returnListOfListOfIntArg: () =>
           new MockList(
             2,
-            (o: any, a: { [key: string]: any }) => new MockList(a['l']),
+            (_o: any, a: { [key: string]: any }) => new MockList(a['l']),
           ),
       }),
       Int: () => 12,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `{
       returnListOfListOfIntArg(l: 1)
     }`;
     const expected = {
       returnListOfListOfIntArg: [[12], [12]],
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1273,16 +1282,16 @@ describe('Mock', () => {
     // unintuitive corner-cases
     const mockMap = {
       RootQuery: () => ({
-        thread: (o: any, a: { [key: string]: any }) => ({ id: a['id'] }),
-        threads: (o: any, a: { [key: string]: any }) =>
+        thread: (_o: any, a: { [key: string]: any }) => ({ id: a['id'] }),
+        threads: (_o: any, a: { [key: string]: any }) =>
           new MockList(ITEMS_PER_PAGE * a['num']),
       }),
       Thread: () => ({
         name: 'Lorem Ipsum',
-        posts: (o: any, a: { [key: string]: any }) =>
+        posts: (_o: any, a: { [key: string]: any }) =>
           new MockList(
             ITEMS_PER_PAGE * a['num'],
-            (oi: any, ai: { [key: string]: any }) => ({ id: ai['num'] }),
+            (_oi: any, ai: { [key: string]: any }) => ({ id: ai['num'] }),
           ),
       }),
       Post: () => ({
@@ -1291,7 +1300,7 @@ describe('Mock', () => {
       }),
       Int: () => 123,
     };
-    addMockFunctionsToSchema({ schema: jsSchema, mocks: mockMap });
+    addMocksToSchema({ schema: jsSchema, mocks: mockMap });
     const testQuery = `query abc{
       thread(id: "67"){
         id
@@ -1314,22 +1323,22 @@ describe('Mock', () => {
         ],
       },
     };
-    return graphql(jsSchema, testQuery).then(res => {
+    return graphql(jsSchema, testQuery).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
 
   it('works for resolvers returning javascript Dates', () => {
     const typeDefs = `
-    	scalar Date
+      scalar Date
 
       type DateObject {
         start: Date!
       }
 
       type Query {
-      	date1: DateObject
-    		date2: Date
+        date1: DateObject
+        date2: Date
         date3: Date
       }
     `;
@@ -1356,7 +1365,7 @@ describe('Mock', () => {
       resolvers,
     });
 
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema,
       mocks: {
         Date: () => new Date('2016-05-04'),
@@ -1381,7 +1390,7 @@ describe('Mock', () => {
       date2: '2016-01-01T00:00:00.000Z',
       date3: '2016-05-04T00:00:00.000Z',
     };
-    return graphql(schema, query).then(res => {
+    return graphql(schema, query).then((res) => {
       expect(res.data).to.deep.equal(expected);
     });
   });
@@ -1398,9 +1407,9 @@ describe('Mock', () => {
     }
 
     const typeDefs = `
-    	interface Node {
-    		id: ID!
-    	}
+      interface Node {
+        id: ID!
+      }
 
       type Account implements Node {
         id: ID!
@@ -1418,17 +1427,15 @@ describe('Mock', () => {
 
     const resolvers = {
       Query: {
-        node: () => {
-          return new Account();
-        },
+        node: () => new Account(),
       },
       Node: {
         __resolveType: (obj: any) => {
           if (obj instanceof Account) {
             return 'Account';
-          } else {
-            return null;
           }
+
+          return null;
         },
       },
     };
@@ -1438,7 +1445,7 @@ describe('Mock', () => {
       resolvers,
     });
 
-    addMockFunctionsToSchema({
+    addMocksToSchema({
       schema,
       preserveResolvers: true,
     });
@@ -1462,7 +1469,7 @@ describe('Mock', () => {
         },
       },
     };
-    return graphql(schema, query).then(res => {
+    return graphql(schema, query).then((res) => {
       expect(res).to.deep.equal(expected);
     });
   });
